requirements.txt
MouseInfo==0.1.3
mss==9.0.1
numpy==1.26.4
opencv-python==4.9.0.80
pandas==2.2.1
pillow==10.3.0
PyAutoGUI==0.9.54
pygame==2.5.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.1
rubicon-objc==0.4.7
six==1.16.0
tzdata==2024.1
